logic fix for override False but symlink already exists

+++ src/murfey/server/api/file_io_frontend.py
@@ -50,10 +50,9 @@ async def create_symlink(
         instrument_name
     ]
     symlink_full_path = machine_config.rsync_basepath / symlink_params.symlink
-    if symlink_full_path.is_symlink():
-        if symlink_params.override:
-            symlink_full_path.unlink()
-    elif symlink_full_path.exists():
+    if symlink_full_path.is_symlink() and symlink_params.override:
+        symlink_full_path.unlink()
+    if symlink_full_path.exists():
         return ""
     symlink_full_path.symlink_to(symlink_params.target)
     return str(symlink_params.symlink)
